+++ PaddleAudio/paddleaudio/utils/__init__.py
@@ -12,9 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# from .download import *
-# from .env import *
-# from .error import *
-# #from .logging import *
-# from .time import *
-from .utils import ParameterError
+from . import _librosa
+from .utils import *
