+++ PaddleAudio/paddleaudio/utils/utils.py
@@ -12,16 +12,68 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+__all__ = [
+    'ParameterError',
+    'get_logger',
+    'Timer',
+    'seconds_to_hms',
+    'download_and_decompress',
+    'load_state_dict_from_url',
+    'default_logger',
+]
+
+import logging
 import math
 import os
+import sys
 import time
+from typing import Dict, List, Optional
 
-import numpy as np
+from paddle.framework import load as load_state_dict
+from paddle.utils import download
 
 
 class ParameterError(Exception):
     """Exception class for Parameter checking"""
     pass
+
+
+def get_logger(name: Optional[str] = None,
+               use_error_log: bool = False,
+               log_dir: Optional[os.PathLike] = None,
+               log_file_name: Optional[str] = None):
+
+    if name is None:
+        name = __file__
+
+    logger = logging.getLogger(name)
+    logging_level = getattr(logging, 'INFO')
+    logger.setLevel(logging_level)
+
+    formatter = logging.Formatter(
+        fmt='%(asctime)s %(filename)s: %(levelname)s: %(message)s')
+    stdout_handler = logging.StreamHandler(sys.stdout)
+    stdout_handler.setLevel(logging_level)
+    stdout_handler.setFormatter(formatter)
+    logger.addHandler(stdout_handler)
+
+    if log_dir:  #logging to file
+        if log_file_name is None:
+            log_file_name = 'log'
+        log_file = os.path.join(log_dir, log_file_name + '.txt')
+        fh = logging.FileHandler(log_file)
+        fh.setLevel(logging_level)
+        fh.setFormatter(formatter)
+        logger.addHandler(fh)
+
+    if use_error_log:
+        stderr_handler = logging.StreamHandler(sys.stderr)
+        stderr_handler.setLevel(logging.WARNING)
+        stderr_handler.setFormatter(formatter)
+        logger.addHandler(stderr_handler)
+
+    logger.propagate = False
+    return logger
 
 
 class Timer(object):
@@ -73,3 +125,32 @@ def seconds_to_hms(seconds: int) -> str:
     s = int(seconds - h * 3600 - m * 60)
     hms_str = '{:0>2}:{:0>2}:{:0>2}'.format(h, m, s)
     return hms_str
+
+
+def download_and_decompress(archives: List[Dict[str, str]], path: str):
+    """
+    Download archieves and decompress to specific path.
+    """
+    if not os.path.isdir(path):
+        os.makedirs(path)
+
+    for archive in archives:
+        assert 'url' in archive and 'md5' in archive, \
+            'Dictionary keys of "url" and "md5" are required in the archive, but got: {list(archieve.keys())}'
+
+        download.get_path_from_url(archive['url'], path, archive['md5'])
+
+
+def load_state_dict_from_url(url: str, path: str, md5: str = None):
+    """
+    Download and load a state dict from url
+    """
+    if not os.path.isdir(path):
+        os.makedirs(path)
+
+    download.get_path_from_url(url, path, md5)
+    return load_state_dict(os.path.join(path, os.path.basename(url)))
+
+
+default_logger = get_logger(__file__)
+download.logger = default_logger
